ui_tests/requirements.txt
allure-pytest==2.10.0
furl==2.1.3
PyHamcrest==2.0.4
pytest==7.1.3
selenium==4.4.3
webdriver-manager==3.8.3